requirements.txt
# OCR and Image Processing
paddleocr>=2.7.0
paddlepaddle>=2.5.0
opencv-python>=4.8.0
Pillow>=10.0.0

# OpenAI Integration
openai>=1.3.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Optional: Barcode scanning
python-barcode>=0.15.1
pyzbar>=0.1.9

# Date parsing
python-dateutil>=2.8.2
